refactor: [#1221] move DB setup to databases context
