+++ src/bootstrap/app.rs
@@ -27,8 +27,9 @@ use crate::core::authentication::handler::KeysHandler;
 use crate::core::authentication::key::repository::in_memory::InMemoryKeyRepository;
 use crate::core::authentication::key::repository::persisted::DatabaseKeyRepository;
 use crate::core::authentication::service;
+use crate::core::databases::setup::initialize_database;
 use crate::core::scrape_handler::ScrapeHandler;
-use crate::core::services::{initialize_database, initialize_whitelist_manager};
+use crate::core::services::initialize_whitelist_manager;
 use crate::core::statistics;
 use crate::core::torrent::manager::TorrentsManager;
 use crate::core::torrent::repository::in_memory::InMemoryTorrentRepository;
