+++ src/servers/udp/handlers.rs
@@ -475,8 +475,8 @@ mod tests {
 
     use super::gen_remote_fingerprint;
     use crate::core::announce_handler::AnnounceHandler;
+    use crate::core::databases::setup::initialize_database;
     use crate::core::scrape_handler::ScrapeHandler;
-    use crate::core::services::initialize_database;
     use crate::core::statistics::event::sender::Sender;
     use crate::core::torrent::repository::in_memory::InMemoryTorrentRepository;
     use crate::core::torrent::repository::persisted::DatabasePersistentTorrentRepository;
@@ -1393,7 +1393,7 @@ mod tests {
                 use mockall::predicate::eq;
 
                 use crate::core::announce_handler::AnnounceHandler;
-                use crate::core::services::initialize_database;
+                use crate::core::databases::setup::initialize_database;
                 use crate::core::statistics;
                 use crate::core::torrent::repository::in_memory::InMemoryTorrentRepository;
                 use crate::core::torrent::repository::persisted::DatabasePersistentTorrentRepository;
